Remove default risk_tolerance from user
When users had a default risk_tolerance of 0, distinguishing between
one with no RiskProfile and one with a very low risk tolerance would
have required a database call (to see if there were any RiskProfile
records associated with them). With default risk_tolerance set to nil,
this can be avoided.