Remove initial deposit field from account#new view
The initial deposit field was causing a quirk whereby money from the
initial deposit wouldn't be invested until another deposit was made. I
tried fixing this with a callback but that broke a number of other
things in testing. I'll investigate later – moving on to other aspects
of the app for now